# Conflicts:
#	.gitignore
#	sample-applications/sample-command-processor-axonserver-client/src/main/resources/application.properties

+++ axonserver-enterprise/src/main/java/io/axoniq/axonserver/enterprise/cluster/snapshot/ApplicationSnapshotDataStore.java
@@ -1,19 +1,22 @@
 package io.axoniq.axonserver.enterprise.cluster.snapshot;
 
 import com.google.protobuf.InvalidProtocolBufferException;
+import io.axoniq.axonserver.access.application.ApplicationController;
+import io.axoniq.axonserver.access.application.JpaApplication;
 import io.axoniq.axonserver.cluster.snapshot.SnapshotDeserializationException;
-import io.axoniq.axonserver.grpc.ProtoConverter;
+import io.axoniq.axonserver.cluster.snapshot.SnapshotContext;
+import io.axoniq.axonserver.grpc.ApplicationProtoConverter;
 import io.axoniq.axonserver.grpc.cluster.SerializedObject;
-import io.axoniq.platform.application.ApplicationController;
-import io.axoniq.platform.application.jpa.Application;
 import reactor.core.publisher.Flux;
 
 import java.util.List;
 
-import static io.axoniq.axonserver.grpc.ProtoConverter.createJpaApplication;
+import static io.axoniq.axonserver.RaftAdminGroup.isAdmin;
+import static io.axoniq.axonserver.grpc.ApplicationProtoConverter.createJpaApplication;
+
 
 /**
- * Snapshot data store for {@link Application} data.
+ * Snapshot data store for {@link JpaApplication} data. This data is only distributed in the _admin context.
  *
  * @author Milan Savic
  * @since 4.1
@@ -22,9 +25,10 @@ public class ApplicationSnapshotDataStore implements SnapshotDataStore {
 
     private final String context;
     private final ApplicationController applicationController;
+    private final boolean adminContext;
 
     /**
-     * Creates Application Snapshot Data Store for streaming/applying {@link Application} data.
+     * Creates JpaApplication Snapshot Data Store for streaming/applying {@link JpaApplication} data.
      *
      * @param context               the application context
      * @param applicationController the application controller used for retrieving/saving applications
@@ -32,6 +36,7 @@ public class ApplicationSnapshotDataStore implements SnapshotDataStore {
     public ApplicationSnapshotDataStore(String context, ApplicationController applicationController) {
         this.context = context;
         this.applicationController = applicationController;
+        this.adminContext = isAdmin(context);
     }
 
     @Override
@@ -40,26 +45,28 @@ public class ApplicationSnapshotDataStore implements SnapshotDataStore {
     }
 
     @Override
-    public Flux<SerializedObject> streamSnapshotData(long fromEventSequence, long toEventSequence) {
-        List<Application> applications = applicationController.getApplicationsForContext(context);
+    public Flux<SerializedObject> streamSnapshotData(SnapshotContext installationContext) {
+        if(! adminContext) return Flux.empty();
+        List<JpaApplication> applications = applicationController.getApplicationsForContext(context);
 
         return Flux.fromIterable(applications)
-                   .map(ProtoConverter::createApplication)
+                   .map(ApplicationProtoConverter::createApplication)
                    .map(this::toSerializedObject);
     }
 
     @Override
     public boolean canApplySnapshotData(String type) {
-        return Application.class.getName().equals(type);
+        return adminContext && JpaApplication.class.getName().equals(type);
     }
+
 
     @Override
     public void applySnapshotData(SerializedObject serializedObject) {
         try {
             io.axoniq.axonserver.grpc.internal.Application applicationMessage = io.axoniq.axonserver.grpc.internal.Application
                     .parseFrom(serializedObject.getData());
-            Application applicationEntity = createJpaApplication(applicationMessage);
-            applicationController.mergeContext(applicationEntity, context);
+            JpaApplication applicationEntity = createJpaApplication(applicationMessage);
+            applicationController.synchronize(applicationEntity);
         } catch (InvalidProtocolBufferException e) {
             throw new SnapshotDeserializationException("Unable to deserialize application data.", e);
         }
@@ -67,12 +74,14 @@ public class ApplicationSnapshotDataStore implements SnapshotDataStore {
 
     @Override
     public void clear() {
-        applicationController.deleteByContext(context);
+        if( adminContext) {
+            applicationController.deleteByContext(context);
+        }
     }
 
     private SerializedObject toSerializedObject(io.axoniq.axonserver.grpc.internal.Application application) {
         return SerializedObject.newBuilder()
-                               .setType(Application.class.getName())
+                               .setType(JpaApplication.class.getName())
                                .setData(application.toByteString())
                                .build();
     }
