+++ axonserver-enterprise/src/main/java/io/axoniq/axonserver/enterprise/cluster/snapshot/UserSnapshotDataStore.java
@@ -1,14 +1,16 @@
 package io.axoniq.axonserver.enterprise.cluster.snapshot;
 
 import com.google.protobuf.InvalidProtocolBufferException;
+import io.axoniq.axonserver.access.jpa.User;
+import io.axoniq.axonserver.access.user.UserRepository;
 import io.axoniq.axonserver.cluster.snapshot.SnapshotDeserializationException;
-import io.axoniq.axonserver.grpc.ProtoConverter;
+import io.axoniq.axonserver.cluster.snapshot.SnapshotContext;
+import io.axoniq.axonserver.grpc.UserProtoConverter;
 import io.axoniq.axonserver.grpc.cluster.SerializedObject;
-import io.axoniq.platform.user.User;
-import io.axoniq.platform.user.UserRepository;
 import reactor.core.publisher.Flux;
 
-import static io.axoniq.axonserver.grpc.ProtoConverter.createJpaUser;
+import static io.axoniq.axonserver.RaftAdminGroup.isAdmin;
+
 
 /**
  * Snapshot data store for {@link User} data.
@@ -18,15 +20,18 @@ import static io.axoniq.axonserver.grpc.ProtoConverter.createJpaUser;
  */
 public class UserSnapshotDataStore implements SnapshotDataStore {
 
+    private static final String TYPE = User.class.getName();
     private final UserRepository userRepository;
+    private final boolean adminContext;
 
     /**
      * Creates User Snapshot Data Store for streaming/applying user data.
      *
      * @param userRepository the repository for retrieving/saving users
      */
-    public UserSnapshotDataStore(UserRepository userRepository) {
+    public UserSnapshotDataStore(String context, UserRepository userRepository) {
         this.userRepository = userRepository;
+        this.adminContext = isAdmin(context);
     }
 
     @Override
@@ -35,15 +40,16 @@ public class UserSnapshotDataStore implements SnapshotDataStore {
     }
 
     @Override
-    public Flux<SerializedObject> streamSnapshotData(long fromEventSequence, long toEventSequence) {
+    public Flux<SerializedObject> streamSnapshotData(SnapshotContext installationContext) {
+        if( ! adminContext) return Flux.empty();
         return Flux.fromIterable(userRepository.findAll())
-                   .map(ProtoConverter::createUser)
+                   .map(UserProtoConverter::createUser)
                    .map(this::toSerializedObject);
     }
 
     @Override
     public boolean canApplySnapshotData(String type) {
-        return User.class.getName().equals(type);
+        return adminContext && TYPE.equals(type);
     }
 
     @Override
@@ -51,7 +57,7 @@ public class UserSnapshotDataStore implements SnapshotDataStore {
         try {
             io.axoniq.axonserver.grpc.internal.User userMessage = io.axoniq.axonserver.grpc.internal.User.parseFrom(
                     serializedObject.getData());
-            User userEntity = createJpaUser(userMessage);
+            User userEntity = UserProtoConverter.createJpaUser(userMessage);
             userRepository.save(userEntity);
         } catch (InvalidProtocolBufferException e) {
             throw new SnapshotDeserializationException("Unable to deserialize user data.", e);
@@ -60,13 +66,14 @@ public class UserSnapshotDataStore implements SnapshotDataStore {
 
     @Override
     public void clear() {
-        // TODO: 1/10/2019 do we create users per context?
-        userRepository.deleteAll();
+        if( adminContext) {
+            userRepository.deleteAll();
+        }
     }
 
     private SerializedObject toSerializedObject(io.axoniq.axonserver.grpc.internal.User user) {
         return SerializedObject.newBuilder()
-                               .setType(User.class.getName())
+                               .setType(TYPE)
                                .setData(user.toByteString())
                                .build();
     }
